refactor: cmd 配置抽取 + 引入 "full" 构建配置
将 cmd 层的 provider 配置、默认值、IOA 运行时函数抽取到独立文件,
同时引入 "full" build tag 将 katana/passive/uncover 移至可选编译:

cmd 配置重组:
- provider_config.go: providerConfig/visionProviderConfig 等
- option_defaults.go: resolveString/defaultInt/applyDefaults 等
- ioa_runtime.go: defaultIOANodeName
- providerConfig 不再直接推断 provider, 留给 provider.Resolve

"full" 构建配置:
- scanner_commands.go / _full.go: 按 build tag 动态注册子命令
- recon_options.go / _stub.go: full 构建暴露 recon 参数, mini 构建隐藏
- scanner_flags_full.go: katana/passive usage 注入
- register_command.go 加 !full tag, register_command_full.go 包含完整注册
- uncover/passive/katana 相关代码加 full tag
- set_uncover_stub.go / uncover_stub.go: 非 full 构建的空实现
- 移除 ioa_store_sqlite.go (sqlite tag 仅用于 ioa server 二进制)
- 移除 IOAOptions.IOADB 字段
- skills/embed.go 增加 skillAvailable 过滤
- .goreleaser.yml 增加 aiscan-full 构建目标
- build.sh 增加 --profile mini/full 选项
- 新增 skills/swarm/SKILL.md
