+++ cmd/ioa_store_memory.go
@@ -1,5 +1,3 @@
-//go:build !sqlite
-
 package cmd
 
 import (
@@ -8,6 +6,5 @@ import (
 )
 
 func openIOAStore(option *Option, logger telemetry.Logger) (ioaserver.Store, func() error, error) {
-	logger.Warnf("ioa_server store=memory: --ioa-db=%q ignored, all state will be lost on restart (rebuild with -tags sqlite to enable persistence)", option.IOADB)
 	return nil, nil, nil
 }
