+++ pkg/e2e/features_full_test.go
@@ -1,0 +1,9 @@
+//go:build e2e && full
+
+package e2e
+
+func buildTags() string { return "emptytemplates noembed full" }
+
+func scannerHelpCommands() []string {
+	return []string{"gogo", "spray", "katana", "zombie", "neutron", "passive", "scan"}
+}
