+++ pkg/tools/integration_test.go
@@ -1,7 +1,7 @@
-//go:build integration
+//go:build full && integration
 
 //	Run with: AISCAN_INTEGRATION=1 FOFA_EMAIL=... FOFA_KEY=... \
-//	  go test -tags integration ./pkg/tools/... -run TestIntegration -v
+//	  go test -tags 'full integration' ./pkg/tools/... -run TestIntegration -v
 package tools
 
 import (
