+++ pkg/tools/passive/passive_test.go
@@ -1,3 +1,5 @@
+//go:build full
+
 package passive
 
 import (
