+++ pkg/tools/register_command_full.go
@@ -1,4 +1,4 @@
-//go:build !full
+//go:build full
 
 package tools
 
@@ -6,7 +6,9 @@ import (
 	"github.com/chainreactors/aiscan/pkg/command"
 	"github.com/chainreactors/aiscan/pkg/telemetry"
 	gogocmd "github.com/chainreactors/aiscan/pkg/tools/gogo"
+	katanacmd "github.com/chainreactors/aiscan/pkg/tools/katana"
 	neutroncmd "github.com/chainreactors/aiscan/pkg/tools/neutron"
+	passivecmd "github.com/chainreactors/aiscan/pkg/tools/passive"
 	"github.com/chainreactors/aiscan/pkg/tools/scan"
 	"github.com/chainreactors/aiscan/pkg/tools/scan/engine"
 	spraycmd "github.com/chainreactors/aiscan/pkg/tools/spray"
@@ -17,6 +19,8 @@ func init() {
 	command.RegisterFactory(command.Factory{
 		Group: "scanner",
 		Build: func(deps *command.Deps, reg *command.CommandRegistry) {
+			reg.Register(katanacmd.New(), "scanner")
+
 			es, _ := deps.EngineSet.(*engine.Set)
 			if es == nil {
 				return
@@ -26,6 +30,10 @@ func init() {
 				logger = telemetry.NopLogger()
 			}
 			proxy := deps.ScannerProxy
+
+			if es.Uncover != nil {
+				reg.Register(passivecmd.New(es.Uncover).WithLogger(logger), "scanner")
+			}
 
 			var scanOpts []scan.Option
 			for _, o := range deps.ScanOpts {
