+++ pkg/tools/scan/engine/recon_test.go
@@ -1,3 +1,5 @@
+//go:build full
+
 package engine
 
 import "testing"
