+++ pkg/tools/scan/engine/set.go
@@ -168,44 +168,3 @@ func initWithCapacity(ctx context.Context, opts resources.Options, caps Capacity
 	set.Capacity = caps
 	return set, nil
 }
-
-// SetupUncover 在 InitWithOptions 之后追加 uncover engine 初始化, 把 ReconOptions
-// 的凭证注入。任一 source 凭证非空就 init; 都为空时 Uncover 留 nil 让 passive 工具不注册。
-// 可被反复调用: opts 的非零字段累加进 e.Recon, engine 基于合并后的全集重建。
-func (e *Set) SetupUncover(opts ReconOptions, logger telemetry.Logger) {
-	if logger == nil {
-		logger = telemetry.NopLogger()
-	}
-	e.Recon = mergeReconOptions(e.Recon, opts)
-	eng := NewUncoverEngine(e.Recon, logger)
-	if len(eng.Sources()) == 0 {
-		return
-	}
-	if e.Uncover != nil {
-		_ = e.Uncover.Close()
-	}
-	e.Uncover = eng
-	logger.Infof("engine=uncover status=ready sources=%v", e.Uncover.Sources())
-}
-
-func mergeReconOptions(base, next ReconOptions) ReconOptions {
-	if next.FofaEmail != "" {
-		base.FofaEmail = next.FofaEmail
-	}
-	if next.FofaKey != "" {
-		base.FofaKey = next.FofaKey
-	}
-	if next.HunterToken != "" {
-		base.HunterToken = next.HunterToken
-	}
-	if next.HunterAPIKey != "" {
-		base.HunterAPIKey = next.HunterAPIKey
-	}
-	if next.IngressProxy != "" {
-		base.IngressProxy = next.IngressProxy
-	}
-	if next.Limit != 0 {
-		base.Limit = next.Limit
-	}
-	return base
-}
