+++ pkg/tools/scan/engine/set_uncover_stub.go
@@ -1,0 +1,7 @@
+//go:build !full
+
+package engine
+
+import "github.com/chainreactors/aiscan/pkg/telemetry"
+
+func (e *Set) SetupUncover(_ ReconOptions, _ telemetry.Logger) {}
