+++ pkg/tools/scan/engine/uncover.go
@@ -1,3 +1,5 @@
+//go:build full
+
 package engine
 
 import (
