+++ pkg/tools/scan/engine/uncover_agents.go
@@ -1,3 +1,5 @@
+//go:build full
+
 package engine
 
 import (
