+++ pkg/tools/scan/engine/uncover_stub.go
@@ -1,0 +1,8 @@
+//go:build !full
+
+package engine
+
+type UncoverEngine struct{}
+
+func (e *UncoverEngine) Sources() []string { return nil }
+func (e *UncoverEngine) Close() error      { return nil }
