+++ skills/availability_full.go
@@ -1,0 +1,5 @@
+//go:build full
+
+package skills
+
+func skillAvailable(string) bool { return true }
